+++ examples/test/juice-machine.test.tsx
@@ -5,17 +5,17 @@ import { App } from "../src/juice-machine/index.js";
 test("juice machine should dispense the users selected juice after payment", async () => {
   const performer = new Performer(<App />);
   performer.start();
-  await performer.waitUntilSettled();
+  await performer.waitUntilListening();
   performer.input({
     role: "user",
     content: [
       { type: "text", text: "I would like Apple and Mango juice please!" },
     ],
   });
-  await performer.waitUntilSettled();
+  await performer.waitUntilListening();
   performer.input({
     role: "user",
     content: [{ type: "text", text: "**insert payment**" }],
   });
-  await performer.waitUntilSettled();
+  await performer.waitUntilListening();
 }, 60_000);