+++ examples/test/open-ended.test.tsx
@@ -5,12 +5,12 @@ import { App } from "../src/open-ended/index.js";
 test("should have chat until ended", async () => {
   const performer = new Performer(<App />);
   performer.start();
-  await performer.waitUntilSettled();
+  await performer.waitUntilListening();
   performer.input({
     role: "user",
     content: "Write a haiku about Javascript frameworks",
   });
-  await performer.waitUntilSettled();
+  await performer.waitUntilListening();
   performer.input({
     role: "user",
     content: "Beautiful! That's all for now. Good bye!",