+++ packages/performer-core/src/render.ts
@@ -54,11 +54,35 @@ type PausedOp = {
   type: "PAUSED";
 };
 
-export type RenderOp = CreateOp | ResumeOp | PausedOp;
+// like paused by explicitly marked as requiring external input to continue
+type ListeningOp = {
+  type: "LISTENING";
+};
+
+type AfterChildrenOp = {
+  type: "AFTER_CHILDREN";
+  payload: {
+    node: PerformerNode;
+  };
+};
+
+export type RenderOp =
+  | CreateOp
+  | ResumeOp
+  | PausedOp
+  | AfterChildrenOp
+  | ListeningOp;
+
+function onlyListening(ops: Record<string, RenderOp>) {
+  return Object.values(ops).every((op) => op.type === "LISTENING");
+}
+
+function noOps(ops: Record<string, RenderOp>) {
+  return Object.keys(ops).length === 0;
+}
 
 export async function render(performer: Performer) {
-  performer.setState("running");
-  logger.debug("call=render");
+  logger.withTag("render").debug("start");
   try {
     const ops = evaluateRenderOps(
       "root",
@@ -75,14 +99,22 @@ export async function render(performer: Performer) {
           continue;
         case "RESUME":
           await performOp(performer, op);
+          continue;
+        case "AFTER_CHILDREN":
+          op.payload.node.hooks.afterChildren!();
+          // ensure that render is queue at least once if afterChildren has no effect
+          performer.queueRender("after children effect");
       }
     }
-    // todo User is consider paused so its not resolving to settled here
-    if (Object.keys(ops).length === 0 && !performer.renderQueued) {
-      performer.setState("settled");
+    if (noOps(ops) && !performer.renderQueued) {
+      performer.setFinished();
+    } else if (onlyListening(ops) && !performer.renderQueued) {
+      performer.setListening();
     }
   } catch (error) {
     performer.onError("root", error);
+  } finally {
+    logger.withTag("render").debug("finally");
   }
 }
 
@@ -125,6 +157,10 @@ export function evaluateRenderOps(
     };
   }
 
+  if (node.status === "LISTENING") {
+    return { [threadId]: { type: "LISTENING" } satisfies ListeningOp };
+  }
+
   if (node.status === "PAUSED") {
     return { [threadId]: { type: "PAUSED" } satisfies PausedOp };
   }
@@ -144,8 +180,16 @@ export function evaluateRenderOps(
       childPrevSibling,
     );
     Object.assign(ops, childOps);
-    if (childThreadId in childOps) {
+    // increment childRenderCount if render op
+    if (
+      Object.values(childOps).find(
+        (op) => op.type === "CREATE" || op.type === "RESUME",
+      )
+    ) {
       node.childRenderCount += 1;
+    }
+    // return if op for current thread otherwise continue
+    if (childThreadId in childOps) {
       return ops;
     }
     if (childPrevSibling) {
@@ -162,12 +206,13 @@ export function evaluateRenderOps(
     freeNode(childNode, node, true);
   }
 
-  if (node.childRenderCount > 0) {
-    if (node.hooks.afterChildren) {
-      node.hooks.afterChildren();
-    }
-    node.childRenderCount = 0;
+  if (node.childRenderCount > 0 && node.hooks.afterChildren) {
+    ops[threadId] = {
+      type: "AFTER_CHILDREN",
+      payload: { node },
+    };
   }
+  node.childRenderCount = 0;
 
   return ops;
 }
@@ -232,7 +277,11 @@ async function renderComponent(performer: Performer, node: PerformerNode) {
     abortController: performer.abortController,
   });
   try {
-    view = node.type(node.props);
+    try {
+      view = node.type(node.props);
+    } finally {
+      clearRenderScope();
+    }
     if (typeof view !== "function") {
       const returnType = view instanceof Promise ? "Promise" : typeof view;
       throw Error(
@@ -240,8 +289,8 @@ async function renderComponent(performer: Performer, node: PerformerNode) {
           `To make async calls in your component use the \`useResource\` hook`,
       );
     }
-    registerView(performer, node, view);
     node.status = "RESOLVED";
+    registerView(performer, node, view);
   } catch (e) {
     if (e instanceof DeferResource) {
       node.status = "PAUSED";
@@ -253,14 +302,13 @@ async function renderComponent(performer: Performer, node: PerformerNode) {
         })
         .catch((error) => performer.onError(node.threadId, error));
     } else if (e instanceof DeferInput) {
-      node.status = "PAUSED";
+      node.status = "LISTENING";
       logPaused(node, "resource");
       performer.setInputNode(node);
+      performer.queueRender("set input");
     } else {
       throw e;
     }
-  } finally {
-    clearRenderScope();
   }
 }