+++ packages/performer-core/test/render.test.tsx
@@ -28,7 +28,7 @@ test("should render and resolve intrinsic element", async () => {
   );
   const performer = new Performer(app);
   performer.start();
-  await performer.waitUntilSettled();
+  await performer.waitUntilFinished();
   const messages = resolveMessages(performer.root);
   expect(messages).toHaveLength(3);
   expect(messages[0].role).toEqual("system");
@@ -59,7 +59,7 @@ test("should render view", async () => {
   const app = <AComponent />;
   const performer = new Performer(app);
   performer.start();
-  await performer.waitUntilSettled();
+  await performer.waitUntilFinished();
   await testHydration(performer);
 });
 
@@ -86,7 +86,7 @@ test("should update prop when signal changes", async () => {
   }
   const performer = new Performer(<App />);
   performer.start();
-  await performer.waitUntilSettled();
+  await performer.waitUntilFinished();
   await testHydration(performer);
 });
 
@@ -106,14 +106,14 @@ test("should update and run message actions when state changes", async () => {
   );
   const performer = new Performer(app);
   performer.start();
-  await performer.waitUntilSettled();
+  await performer.waitUntilFinished();
   let messages = resolveMessages(performer.root);
   expect(messages).toHaveLength(1);
 
   performer.root!.child!.nextSibling!.hooks["state-0"].value = true;
   performer.start();
 
-  await performer.waitUntilSettled();
+  await performer.waitUntilFinished();
   messages = resolveMessages(performer.root);
   expect(messages).toHaveLength(3);
   await testHydration(performer);
@@ -139,7 +139,7 @@ test("should update links when elements are reordered", async () => {
   );
   const performer = new Performer(app, { logLevel: "trace" });
   performer.start();
-  await performer.waitUntilSettled();
+  await performer.waitUntilFinished();
 
   let messages = resolveMessages(performer.root, undefined);
   expect(messages).toHaveLength(3);
@@ -148,7 +148,7 @@ test("should update links when elements are reordered", async () => {
   offset!.value += 1;
   performer.start();
 
-  await performer.waitUntilSettled();
+  await performer.waitUntilFinished();
   messages = resolveMessages(performer.root, undefined);
   expect(messages).toHaveLength(3);
   await testHydration(performer);
@@ -166,7 +166,7 @@ test("should render new elements when dynamically added or removed", async () =>
   );
   let performer = new Performer(app, { logLevel: "trace" });
   performer.start();
-  await performer.waitUntilSettled();
+  await performer.waitUntilFinished();
   let messages = resolveMessages(performer.root, undefined);
   expect(messages).toHaveLength(1);
 
@@ -177,7 +177,7 @@ test("should render new elements when dynamically added or removed", async () =>
   times.value += 4;
   performer.start();
   // second run
-  await performer.waitUntilSettled();
+  await performer.waitUntilFinished();
   messages = resolveMessages(performer.root, undefined);
   expect(messages).toHaveLength(5);
 
@@ -188,7 +188,7 @@ test("should render new elements when dynamically added or removed", async () =>
   times.value -= 2;
   performer.start();
   // third run
-  await performer.waitUntilSettled();
+  await performer.waitUntilFinished();
   messages = resolveMessages(performer.root, undefined);
   expect(messages).toHaveLength(3);
 
@@ -199,7 +199,7 @@ test("should render new elements when dynamically added or removed", async () =>
   times.value -= 1;
   performer.start();
   // fourth run
-  await performer.waitUntilSettled();
+  await performer.waitUntilFinished();
   messages = resolveMessages(performer.root, undefined);
   expect(messages).toHaveLength(2);
   // final hydration test
@@ -225,7 +225,7 @@ test("should unlink messages when removed by conditional", async () => {
   );
   const performer = new Performer(app);
   performer.start();
-  await performer.waitUntilSettled();
+  await performer.waitUntilFinished();
   let messages = resolveMessages(performer.root, undefined);
   expect(
     messages.length,
@@ -236,7 +236,7 @@ test("should unlink messages when removed by conditional", async () => {
   predicate.value = false;
   performer.start();
 
-  await performer.waitUntilSettled();
+  await performer.waitUntilFinished();
   messages = resolveMessages(performer.root, undefined);
   expect(
     messages.length,
@@ -263,7 +263,7 @@ test("should wait for async message actions", async () => {
   const performer = new Performer(app);
   console.time("Render");
   performer.start();
-  await performer.waitUntilSettled();
+  await performer.waitUntilFinished();
   console.timeEnd("Render");
   const messages = resolveMessages(performer.root);
   expect(messages).toHaveLength(3);
@@ -310,7 +310,7 @@ test("should render tree", async () => {
   );
   const performer = new Performer(app);
   performer.start();
-  await performer.waitUntilSettled();
+  await performer.waitUntilFinished();
   const root = performer.root!;
   expect(root.parent).toBeUndefined();
   expect(root.nextSibling).toBeUndefined();
@@ -344,7 +344,7 @@ test("should catch sync component that throws", async () => {
   performer.addEventListener("error", (event) => {
     events.push(event);
   });
-  await performer.waitUntilSettled();
+  await performer.waitUntilFinished();
   expect(events).toHaveLength(1);
 });
 
@@ -360,7 +360,7 @@ test("should catch resumed component that throws", async () => {
   performer.addEventListener("error", (event) => {
     errors.push(event);
   });
-  await performer.waitUntilSettled();
+  await performer.waitUntilFinished();
   expect(errors).toHaveLength(1);
 });
 
@@ -371,7 +371,7 @@ test("should throw if component is async function", async () => {
   // @ts-ignore
   const performer = new Performer(<App />, { throwOnError: false });
   performer.start();
-  await performer.waitUntilSettled();
+  await performer.waitUntilFinished();
   expect(performer.errors).toHaveLength(1);
 });
 
@@ -388,7 +388,7 @@ test("should cast non-string message children", async () => {
   }
   const performer = new Performer(<App />);
   performer.start();
-  await performer.waitUntilSettled();
+  await performer.waitUntilFinished();
   const messages = resolveMessages(performer.root);
   expect(messages[0].content).toEqual(
     "Message with 1 and 0 and true and false and null and undefined [object Object]",