+++ packages/performer-playground/src/components/ChatWindow.tsx
@@ -5,9 +5,13 @@ import { Tabs, TabsContent, TabsList, TabsTrigger } from "./ui/tabs.js";
 import { MessageList } from "./MessageList.js";
 import { useScroll } from "../hooks/use-scroll.js";
 import { createContext, useEffect } from "react";
+import { PerformerOptions } from "@performer/core";
 
 export function ChatWindow({ App }: { App: Component<any> }) {
-  const { events, sendMessage } = usePerformerClient(App);
+  const options: PerformerOptions = {
+    logLevel: import.meta.env["VITE_LOGLEVEL"] || "info",
+  };
+  const { events, sendMessage, state } = usePerformerClient(App, options);
 
   const {
     messagesStartRef,
@@ -55,6 +59,7 @@ export function ChatWindow({ App }: { App: Component<any> }) {
       </div>
       <div className="w-full pt-2 dark:border-white/20 md:w-[calc(100%-.5rem)] md:border-transparent md:pt-0 md:dark:border-transparent">
         <MessageInput
+          state={state}
           disclaimer="AI can make mistakes. Consider checking important information."
           onSubmit={(text) => sendMessage(text)}
         />