+++ packages/performer-playground/src/components/MessageInput.tsx
@@ -1,15 +1,16 @@
 import { KeyboardEventHandler, useEffect, useRef, useState } from "react";
+import { PerformerState } from "@performer/core";
 
 export type MessageInputProps = {
   disclaimer?: string;
-  disabled?: boolean;
   onSubmit: (text: string) => void;
+  state: PerformerState;
 };
 
 export function MessageInput({
   disclaimer,
-  disabled = false,
   onSubmit,
+  state,
 }: MessageInputProps) {
   const [text, setText] = useState("");
   const textAreaRef = useRef<HTMLTextAreaElement>(null);
@@ -35,6 +36,12 @@ export function MessageInput({
     }
   };
 
+  useEffect(() => {
+    if (textAreaRef.current && state === "listening") {
+      textAreaRef.current.focus();
+    }
+  }, [state]);
+
   return (
     <>
       <form
@@ -52,7 +59,12 @@ export function MessageInput({
                 tabIndex={0}
                 data-id="root"
                 rows={1}
-                placeholder="Message Performer…"
+                disabled={state !== "listening"}
+                placeholder={
+                  state === "finished"
+                    ? "Performer has finished."
+                    : "Message Performer…"
+                }
                 className="m-0 w-full resize-none border-0 bg-transparent py-[10px] pl-3 pr-10 placeholder-black/50 focus:outline-0 focus:ring-0 focus-visible:ring-0 dark:bg-transparent dark:placeholder-white/50 md:py-3.5 md:pl-4 md:pr-12"
                 ref={textAreaRef}
                 value={text}
@@ -62,7 +74,7 @@ export function MessageInput({
               ></textarea>
               <button
                 ref={submitBtnRef}
-                disabled={disabled}
+                disabled={state !== "listening"}
                 className="absolute bottom-1.5 right-2 rounded-lg border border-black p-0.5 text-white transition-colors enabled:bg-black disabled:bg-black disabled:text-gray-400 disabled:opacity-10 dark:border-white dark:bg-white dark:hover:bg-gray-900 dark:disabled:bg-white dark:disabled:hover:bg-transparent md:bottom-3 md:right-3"
                 data-testid="send-button"
               >