+++ examples/CHANGELOG.md
@@ -1,5 +1,15 @@
 # performer-examples
 
+## 0.0.38
+
+### Patch Changes
+
+- dbba314: Add state property to Performer class to replace hasFinished logic
+- Updated dependencies [fc10871]
+- Updated dependencies [dbba314]
+  - @performer/core@0.0.39
+  - @performer/playground@0.0.39
+
 ## 0.0.37
 
 ### Patch Changes